requirements.txt
anyio==3.6.2
arrow==1.2.3
binaryornot==0.4.4
black==22.10.0
certifi==2022.9.24
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.6
cookiecutter==1.7.3
cycler==0.11.0
fastapi==0.79.1
fonttools==4.38.0
h11==0.12.0
httpcore==0.15.0
httptools==0.5.0
httpx==0.23.0
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
kiwisolver==1.4.4
loguru==0.6.0
MarkupSafe==2.1.1
matplotlib==3.6.2
multidict==6.0.2
mypy-extensions==0.4.3
nb-cli==0.6.7
nonebot-adapter-mirai2==0.0.17
nonebot2==2.0.0rc1
numpy==1.23.4
packaging==21.3
pathspec==0.10.1
Pillow==9.3.0
platformdirs==2.5.2
poyo==0.5.0
prompt-toolkit==3.0.31
pydantic==1.9.2
pyfiglet==0.8.post1
pygtrie==2.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-slugify==6.1.2
PyYAML==6.0
requests==2.28.1
rfc3986==1.5.0
scipy==1.9.3
six==1.16.0
sniffio==1.3.0
starlette==0.19.1
text-unidecode==1.3
thefuzz==0.19.0
tomli==2.0.1
tomlkit==0.10.2
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.18.0
wcwidth==0.2.5
websockets==10.4
yarl==1.8.1
